# tpl: Add a partial lookup cache