--- a/resources/page/pages_related.go
+++ b/resources/page/pages_related.go
@@ -124,7 +124,7 @@ func (p Pages) withInvertedIndex(ctx context.Context, search func(idx *related.I
 		return nil, nil
 	}
 
-	d, ok := p[0].(InternalDependencies)
+	d, ok := p[0].(RelatedDocsHandlerProvider)
 	if !ok {
 		return nil, fmt.Errorf("invalid type %T in related search", p[0])
 	}